rfc5321: extract tlsoptions into separate module
