+++ crates/rfc5321/src/tls.rs
@@ -1,0 +1,121 @@
+#![cfg(feature = "client")]
+use hickory_proto::rr::rdata::TLSA;
+use openssl::ssl::SslOptions;
+use std::sync::Arc;
+use tokio_rustls::rustls::crypto::{aws_lc_rs as provider, CryptoProvider};
+use tokio_rustls::rustls::{ClientConfig, SupportedCipherSuite};
+use tokio_rustls::TlsConnector;
+
+#[derive(Debug, Clone, Default)]
+pub struct TlsOptions {
+    pub insecure: bool,
+    pub alt_name: Option<String>,
+    pub dane_tlsa: Vec<TLSA>,
+    pub prefer_openssl: bool,
+    pub openssl_cipher_list: Option<String>,
+    pub openssl_cipher_suites: Option<String>,
+    pub openssl_options: Option<SslOptions>,
+    pub rustls_cipher_suites: Vec<SupportedCipherSuite>,
+}
+
+impl TlsOptions {
+    pub fn build_tls_connector(&self) -> TlsConnector {
+        let cipher_suites = if self.rustls_cipher_suites.is_empty() {
+            provider::DEFAULT_CIPHER_SUITES
+        } else {
+            &self.rustls_cipher_suites
+        };
+
+        let provider = Arc::new(CryptoProvider {
+            cipher_suites: cipher_suites.to_vec(),
+            ..provider::default_provider()
+        });
+
+        let config = ClientConfig::builder_with_provider(provider.clone())
+            .with_protocol_versions(tokio_rustls::rustls::DEFAULT_VERSIONS)
+            .expect("inconsistent cipher-suite/versions selected");
+
+        let config = if self.insecure {
+            config
+                .dangerous()
+                .with_custom_certificate_verifier(Arc::new(danger::NoCertificateVerification::new(
+                    provider.clone(),
+                )))
+        } else {
+            config
+                .dangerous()
+                .with_custom_certificate_verifier(Arc::new(
+                    rustls_platform_verifier::Verifier::new().with_provider(provider),
+                ))
+        };
+        let config = config.with_no_client_auth();
+
+        TlsConnector::from(Arc::new(config))
+    }
+}
+
+mod danger {
+    use std::sync::Arc;
+    use tokio_rustls::rustls::client::danger::{
+        HandshakeSignatureValid, ServerCertVerified, ServerCertVerifier,
+    };
+    use tokio_rustls::rustls::crypto::{
+        verify_tls12_signature, verify_tls13_signature, CryptoProvider,
+    };
+    use tokio_rustls::rustls::pki_types::{CertificateDer, ServerName, UnixTime};
+    use tokio_rustls::rustls::DigitallySignedStruct;
+
+    #[derive(Debug)]
+    pub struct NoCertificateVerification(Arc<CryptoProvider>);
+
+    impl NoCertificateVerification {
+        pub fn new(provider: Arc<CryptoProvider>) -> Self {
+            Self(provider)
+        }
+    }
+
+    impl ServerCertVerifier for NoCertificateVerification {
+        fn verify_server_cert(
+            &self,
+            _end_entity: &CertificateDer<'_>,
+            _intermediates: &[CertificateDer<'_>],
+            _server_name: &ServerName<'_>,
+            _ocsp: &[u8],
+            _now: UnixTime,
+        ) -> Result<ServerCertVerified, tokio_rustls::rustls::Error> {
+            Ok(ServerCertVerified::assertion())
+        }
+
+        fn verify_tls12_signature(
+            &self,
+            message: &[u8],
+            cert: &CertificateDer<'_>,
+            dss: &DigitallySignedStruct,
+        ) -> Result<HandshakeSignatureValid, tokio_rustls::rustls::Error> {
+            verify_tls12_signature(
+                message,
+                cert,
+                dss,
+                &self.0.signature_verification_algorithms,
+            )
+        }
+
+        fn verify_tls13_signature(
+            &self,
+            message: &[u8],
+            cert: &CertificateDer<'_>,
+            dss: &DigitallySignedStruct,
+        ) -> Result<HandshakeSignatureValid, tokio_rustls::rustls::Error> {
+            verify_tls13_signature(
+                message,
+                cert,
+                dss,
+                &self.0.signature_verification_algorithms,
+            )
+        }
+
+        fn supported_verify_schemes(&self) -> Vec<tokio_rustls::rustls::SignatureScheme> {
+            self.0.signature_verification_algorithms.supported_schemes()
+        }
+    }
+}
